pyproject.toml
[tool.poetry]
name = "doctor-front-api"
version = "0.1.0"
description = ""
authors = ["NOOR OMARI <nooromari847@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
pytest = "^6.2.5"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
